Replace some of ESTL with STL
libptty originally used their own implementation of STL called ESTL,
propably to provide support for STL features that weren't implemented in
all platforms and compilers. Replace many of these with usage of STL as
we want to use as much of the standard library as possible, instead of
implementing stuff on our own.